Protocol cleanup: Updated JSON endpoint data and moved to SSL connections.
This is a fairly big deal! The JSON endpoints for kinks and other data were out of date, which made stuff not work (at least in the debug logs). A such, I went head and ported over some changes from the flist-pidgin source. Furthermore, the connections to F-Chat have all been unencrypted up until now. It might not seem like a big deal to you, but I, for one, don't want anyone in the local network to be able to snoop in on my chats. Worry no more! This version moves all connections over to SSL, so all connections to and from F-List will be encrypted. Side note: there is a bug in the current version of Adium that breaks compatibility with many SSL servers, including F-List. For this updated version to work, you either need to wait for the patch at https://trac.adium.im/ticket/16678 to be accepted and released, or patch the Adium source yourself manually and build it. Until that patch is integrated or some other fix is in place, this plugin won't work, and as such, I won't be making a release for this yet.