Some fixes for BBCode and other formatting bugs.
The [sub] tag should work now (or at least it won't be left in). I'm not sure if Adium accepts it as a style, or if it's a theme-specific thing, but mine didn't show the styling for it. Beyond that, there are fixes for sending < in messages, as well as BBCode in ads sent by users.